+++ guess.h
@@ -9,19 +9,18 @@ class Guess {
     char letters[8];
     char guess[8];
     char guess_word[8] = "";
-    int guess_ptr = 0;
+    uint8_t guess_ptr = 0;
 
-    int cursor_pos = 0;
+    uint8_t cursor_pos = 0;
 
     int reset(char * input_letters);
 
-    int cursorMove(int step);
+    int cursorMove(uint8_t step);
 
     int handleSelect();
     int handleDelete();
 
     int handleReturnShuffle();
-
 };
 
 #endif
